requirements.txt
alembic==1.7.5
anyio==3.4.0
asgiref==3.4.1
asyncpg==0.25.0
autopep8==1.6.0
click==8.0.3
databases==0.5.3
fastapi==0.71.0
flake8==4.0.1
greenlet==1.1.2
h11==0.12.0
idna==3.3
isort==5.10.1
Mako==1.1.6
MarkupSafe==2.0.1
mccabe==0.6.1
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
python-dotenv==0.19.2
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.17.1
toml==0.10.2
typing_extensions==4.0.1
uvicorn==0.16.0
